requirements.txt
 openai-whisper
 numpy
 sounddevice
 keyboard
